Fixed format specifiers

+++ skeletons/constr_SET_OF.c
@@ -960,7 +960,7 @@ SET_OF_decode_uper(const asn_codec_ctx_t *opt_codec_ctx,
 		/* X.691, #19.5: No length determinant */
 		nelems = per_get_few_bits(pd, ct->effective_bits);
 		ASN_DEBUG("Preparing to fetch %ld+%ld elements from %s",
-			(long)nelems, ct->lower_bound, td->name);
+			(long)nelems, (long)ct->lower_bound, td->name);
 		if(nelems < 0)  ASN__DECODE_STARVED;
 		nelems += ct->lower_bound;
 	} else {
@@ -1041,7 +1041,7 @@ SET_OF_encode_uper(const asn_TYPE_descriptor_t *td,
     if(ct) {
         int not_in_root =
             (list->count < ct->lower_bound || list->count > ct->upper_bound);
-        ASN_DEBUG("lb %ld ub %ld %s", ct->lower_bound, ct->upper_bound,
+        ASN_DEBUG("lb %jd ub %jd %s", ct->lower_bound, ct->upper_bound,
                   ct->flags & APC_EXTENSIBLE ? "ext" : "fix");
         if(ct->flags & APC_EXTENSIBLE) {
             /* Declare whether size is in extension root */
@@ -1139,7 +1139,7 @@ SET_OF_encode_aper(const asn_TYPE_descriptor_t *td,
     if(ct) {
         int not_in_root =
             (list->count < ct->lower_bound || list->count > ct->upper_bound);
-        ASN_DEBUG("lb %ld ub %ld %s", ct->lower_bound, ct->upper_bound,
+        ASN_DEBUG("lb %jd ub %jd %s", ct->lower_bound, ct->upper_bound,
                   ct->flags & APC_EXTENSIBLE ? "ext" : "fix");
         if(ct->flags & APC_EXTENSIBLE) {
             /* Declare whether size is in extension root */
@@ -1233,7 +1233,7 @@ SET_OF_decode_aper(const asn_codec_ctx_t *opt_codec_ctx,
 		/* X.691, #19.5: No length determinant */
 		nelems = aper_get_nsnnwn(pd, ct->upper_bound - ct->lower_bound + 1);
 		ASN_DEBUG("Preparing to fetch %ld+%ld elements from %s",
-		          (long)nelems, ct->lower_bound, td->name);
+		          (long)nelems, (long)ct->lower_bound, td->name);
 		if(nelems < 0)  ASN__DECODE_STARVED;
 		nelems += ct->lower_bound;
 	} else {
